+++ extensions/builtbybit/CHANGELOG.md
@@ -1,3 +1,7 @@
 # BuiltByBit Changelog
 
 ## [Initial Release] - {PR_MERGE_DATE}
+
+- Search for members based on their username, user ID, or Discord ID.
+- Search for resources by author name.
+- Monitor your notifications from your menu bar!